+++ spec/rspec/core/example_group_spec.rb
@@ -889,7 +889,7 @@ module RSpec::Core
         group.run.should be_true
       end
 
-      it "should raise proper error message when shared context is not found" do
+      it "raises a friendly error message when shared context is not found" do
         expect do
           group = ExampleGroup.describe do
             include_context "not exist"